Доработки и правки, добавление сложного теста с дробными значениями.

+++ WebCalculator/Services/Abstraction/ICalculator.cs
@@ -2,6 +2,6 @@
 {
     public interface ICalculator
     {
-        public double Calculate(string expression);
+        public decimal Calculate(string expression);
     }
 }
