+++ WebCalculator/Services/Abstraction/IOperation.cs
@@ -3,6 +3,6 @@
     public interface IOperation
     {
         int Priority { get; }
-        double Execute(double left, double right);
+        decimal Execute(decimal left, decimal right);
     }
 }
